conda_requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4~=4.12.3
lxml
PyYAML
requests==2.32.3
flask==3.0.3
ollama==0.3.2
pypdf
langchain
langchain-chroma
langchain-ollama
langchain-huggingface
fpdf2
opencv-python
pandas
pillow>=6.2.0
pytesseract>=0.2.6
werkzeug>=2.0
certifi==2018.11.29
Click>=7.1.2
Flask>=2.0.1
gunicorn>=19.9.0
itsdangerous>=2.0
Jinja2>=2.10.1
MarkupSafe>=2.0
pytest==6.2.3
pytest-cov==2.11.1
pylint==2.7.4
langgraph==0.2.60
langgraph-checkpoint-sqlite==2.0.1
sqlalchemy==1.4.41
google-generativeai
fastapi
uvicorn[standard]
jinja2
python-multipart
python-dotenv
pytest-playwright